requirements.txt
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-tavily # For Tavily Search tool
wikipedia
python-dotenv
pydantic
requests # For Relevance AI API and other HTTP requests

# Added from the other requirements list
beautifulsoup4
lxml
httpx
imaplib2 ; sys_platform != "win32"
tenacity
structlog

# Ensure specific versions if needed, e.g.:
# langchain==0.1.16
# pydantic>=2.0 